backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.40
alembic==1.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
joblib==1.5.0
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2024.2
Werkzeug==3.1.3
